refactor: RacingGame 테스트코드를 CarTest로 변경
